requirements.txt
streamlit>=1.35.0
ultralytics==8.3.176
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow>=10.0.0
timm==1.0.14
PyYAML==6.0.1
scipy==1.13.0
psutil==5.9.8
